Use the `jakirkham/centos_conda` image as base
Instead of using the `centos:6` image as the base image, use the
`jakirkham/centos_conda` image as the base image. Doing this as the
new base image already includes working Miniconda2 and Miniconda3
installs. So there is no need to include them here.

Further this will simplify the build process when trying to manage the
`nanshe/nanshe` image with SGE builds and non-SGE builds. This is in
preparation of non-SGE builds of the workflow. The latter being useful
for creating Singularity images from the Docker images, which the
cluster now can use.