Cleanup `.pki` directory after SGE install
Appears that the `yum install` steps used by SGE result in generating
the `.pki` directory in `root`'s home again despite having removed it
before after downloading Miniconda. To make sure the `.pki` directory is
not included in the Docker commit for the SGE install layer, remove it
at the very end of the `install_ge.sh` script.